refined referral handling. next step is hashing user password on registration.
